requirements.txt
# Core dependencies
google-adk==1.14.1
a2a-sdk==0.3.5
fastapi==0.116.2
litellm==1.77.0
uvicorn[standard]==0.35.0
streamlit==1.49.1
python-dotenv==1.1.1
httpx==0.28.1
click